refactor: touch on some style and change design (#1037)
* refactor: touch on some style and change design

* fix: update import path for sample json
