+++ src/pages/index.tsx
@@ -5,9 +5,9 @@ import RenderIf from "@/components/UI/Common/RenderIf";
 import Drawer from "@/components/UI/Drawer/Drawer";
 import FilterMenu from "@/components/UI/FilterMenu/FilterMenu";
 import Footer from "@/components/UI/Footer/Footer";
-import FooterBanner from "@/components/UI/FooterBanner/FooterBanner";
+import FooterBanner from "@/components/UI/Footer/Banner";
 import LocaleSwitch from "@/components/UI/I18n/LocaleSwitch";
-import ScanAreaButton from "@/components/UI/ScanAreaButton/ScanAreaButton";
+import ScanAreaButton from "@/components/UI/Button/ScanArea";
 import SitesIcon from "@/components/UI/SitesIcon/Icons";
 import { MaintenanceError } from "@/errors";
 import { useVerifiedLocations } from "@/hooks/useVerifiedLocations";
